requirements.txt
ftfy==6.3.1
matplotlib==3.5.2
numpy==2.1.3
open_clip_torch==2.22.0
pandas==2.2.3
Pillow==9.2.0
Pillow==11.0.0
python_magic==0.4.27
regex==2023.10.3
Requests==2.32.3
scikit_learn==1.2.2
seaborn==0.11.2
tabulate==0.9.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1